KLibc: regenerate libs
Oops, when I rebuilt the libs there was an error in rpterr.r and it
assembled to zero bytes. Unfortunately, unlike the "normal" shared libs we
use today on modern systems, you can't _really_ tell if you did it right
on OS-9 until you start rebuilding everything with the new lib.

This one actually works. :)