Add some help files for programs we're responsible for
Since we have put a binary for the new Help system in, we should probably
install some damn help files, at least for the stuff we're responsible
for.

This is the help for DCC, Help, Lex, Make, RDump, RLink, RMA, Sed, Touch,
and Yacc. The docs for libc are coming.