COpt: fix a crash/segfault
This is a definite brown-bagger. In my "matched copy" routine, I
failed to initialize the "source pointer" register variable to a
useful value. Most of the time it turned out to be harmless on 6x09,
but that's no way to run a railroad and it blew up on x86 Linux
the first time it encountered a really weird value while copying the
result of a match into a new instruction.