COpt: fix the label bug
I finally found (and fixed) the bug that was deleting labels in
search/replace patterns. In the process, I came to the realization
that we probably should not allow a pattern to be replaced if it
has a label _embedded_ in it, because there's no automatic way to
know which new instruction that label should be moved to.

So, this commit successfully moves "any" labels it sees to the
first instruction of the replacement pattern, but aborts any match
that has one or more labels not attached to the first instruction.

Regenerated dco68 binary.

For GitHub: (Closes #1)