CC09: Implement string merging
Modern C is able to (indeed, is required to) merge two strings that
are next to one another and consider them a single string.
Historically this became necessary when the C preprocessor started
tokenizing and lost the ability to insert text inside strings, but
it also enables there to be really huge strings embedded in
programs without having to use arrays of strings, etc.

That is, this construct:

char *foo = "This is a string\n"
            "covering two lines\n";

should work identically to this construct:

char *foo = "This is a string\ncovering two lines\n";

This is somewhat experimental, but the resulting compiler seems to
work in every case I've thrown at it so far, so it doesn't appear to
break anything.