Header revamp: <math.h>
New, much more documented, math.h header - including more functions
and a global trig variable that had not been documented at all.

Added a more sensible name for the "antilg" function -- exp10() is
the same function, but makes it more clear that it returns 10^x,
where x is a double. It's a companion to exp(), which returns e^x.
Both functions also exist in modern standard libraries, while the
name antilg() is as far as I know unique to our lib.