Revamp doubles
Discovered a number of limits our C library had that were
unnecessarily restrictive.

The first, and most critical, was the scale() function used by
atof() and the compiler. This function essentially shifts the
decimal point of a floating point number. It was so bad that (for
example) the smallest possible double value that could be entered
into either was "3e-39", which has a bit pattern of
"0x02ab1e27ff111401".

The 01 at the end is good, that's the minimum possible exponent for
a positive number. But any bits left in the mantissa (everything
between the top bit and the bottom byte) represent numbers that
can't be reached -- and therefore huge numeric gaps in lots of other
places.

The problem turned out to be the fact that scale() is shifting that
decimal by multiplying or dividing by powers of 10, and some ways of
doing that cause trouble.

Looking at old BSD sources, I noticed that instead of multiplying by
10 a bunch of times, they multiplied by 5 a bunch of times and then
by 2 the same number of times. Recalling that a^x * a^y = a^(xy), I
understood this should have the "same" result, but with less chance
of overflow.

Along the way I implemented versions of the C frexp(), ldexp(), and
scalbn() functions (the last two are actually the same function) for
our floating point format and added them to the standard library.

As it turns out, I didn't need to rewrite atof() itself, just the
scale() function. This one is a little bigger (and I do mean a
little), but it's way more correct -- the smallest double is now
2.938735877055719e-39, which has the bit pattern 0x0000000000000301
and is only 3/(2^56) away from zero.

This is just barely small enough to fit in the compiler, but
shrinking it even further is another project.