CC09: Shrink scale() for the compiler
This is a somewhat dirty version of the C library's new scale() function,
made just to (significantly) reduce its impact on the compiler's size.

This does the same job as libc's version, but it doesn't do anything the
compiler doesn't need. It relies on the compiler's error checking to not
fail, so it happens to save 106 precious code bytes!