Delete some generated files, add others (grr...)
The C library's build system has a few compiled objects it depends on, but
does not build (or include any source for). These include the library
functions ltol3(), l3tol(), qsort(), scanf(), and pffinit().

In my spare time I have been "decompiling" (reverse-engineering a
disassembly into C source that produces the original disassembly) scanf(),
and I think there might be assembly source that is compatible with l3tol
and ltol3 elsewhere, but not linked in...but these are the objects that
Carl actually used to make clib[t].

As I get these projects completed, all of these [non-]generated files
should go away.

And speaking of which, we don't need to keep the intermediate lib files we
make in version control.