KLibc: sleep() -- fix hang on CoCo NitrOS-9 Level 1
I'm not entirely sure why, but the algorithm klibc uses to determine the
number of ticks in a second seems to hang on NitrOS-9 level 1 when running
on a CoCo.

At the same time, we want to kill a few extra birds with this stone. :)

This commit adds a function _stps() that guesses, caches, and returns the
guessed number of clock ticks per second. The first call does the guess,
subsequent calls just return the value.

If you happen to be on a machine that the code guesses the wrong value for,
you can declare "extern int _tps;" in your program and set _tps before
calling sleep() for the first time.

Note: This code uses the SS.Joy "getstat" to detect a Color Computer and
assumes that the machine is 60Hz. If you have a better way to detect a CoCo,
or to detect 50/60Hz on one, I am all ears. Seriously.