CPP: Fix a fix.
As it turns out, I was overzealous in sanitizing the putesc() function's
inputs -- the function's second argument, "arg", may validly be zero. We
were overwriting zeroes with the address of the empty string (and then
interpreting that value as an int). This isn't dangerous with our compiler,
but it does result in the compiler not knowing the line errors occurred on.
Umm, oops?

DCPP and CPP binaries regenerated, and they gain the benefit of the earlier
*brk divorce, losing ~124 bytes each.

This commit partially reverts b97a2bef992148d10ab16252a3589480aeb04091