KLibc: fix u2otime()
The OS-9 date/time format (struct sgtbuf) uses the numbers 1-12 to
represent the months of the year, while C's struct tm uses the numbers 0-11
for the same thing. So any time we convert between them, we need to add or
subtract 1 from the month, depending on the direction.

Michael Furman (@n6il) alerted me that u2otime() doesn't actually do this,
so let's fix that.