<errno.h>: Add a EDOM (domain error), correct ENOMEM
The C standards say that trig functions set errno to EDOM when they fail.
Ours currently use error 44, which has the symbol EILLARG in the header.

Let's make EDOM a synonym for EILLARG; at some point, I may switch that,
or get rid of EILLARG because it's a "C-style" error symbol without a
standard.

Also, the C standards say that sbrk() sets errno to ENOMEM when it fails.
I initially used the E_NORAM (237) OS-9 error number for this, which OS-9
returns when there is no memory at all available anywhere, but that's
not the error code returned by the F$Mem system call -- that's code 207,
"Process Memory Full", returned when the system has free memory, but can't
give any to _you_ for whatever reason.