Defs: Update and add a BUNCH of standard C/Unix headers
These significantly increase our level of compatibility with programs
written for Unix and other systems. Everything I know of that should be
declared is being declared, in the right spots (or in what passes for the
right spots), to give us maximum compatibility and hopefully efficiency.

I have been using these headers for months now to port stuff, and it's
working pretty well.