DLogin: Seems everyone has to write one eventually...
It's almost like a rule or something, everybody seems to make a login
program. Maybe it's because CoCo OS-9 Level 2 shipped with neither Login
nor TSMon, or because it's a program that's simple enough for a novice to
write, but complicated enough to make it a little tricky.

My point with this one was to replace the NitrOS-9 login program, which
doesn't seem to be able to turn off terminal echo (and thus puts your
password onto your screen). But also I wanted to do it while showing how
to 'make your bytes count' in OS-9 C. For example, this program makes very
little use of the standard library -- no strlen, strcpy, or "FILE *" I/O,
so it compiles to a binary just over 3 kilobytes in size. To compare, the
StG login binary was 11K and a bare-bones one I found in an archive was
5K, so it's easy to do worse. :)

Before displaying the login prompt, /DD/SYS/banner is copied to the
terminal, and after a successful login /DD/SYS/motd is as well, before
chaining to the user's shell.

Unlike most OS-9 login programs, any user in the password file that has *
for a password is ineligible to login (empty passwords are still legal).
Failed logins induce delays before retry can be attempted, and there is
no way to tell a bad password from a bad username so there should be less
(I won't say "no") information leakage.