CGfx: Implement _ss_gip2(path,options)
This SetStat call was introduced in NitrOS-9 "Ease of Use" beta 6. The
"options" are a somewhat unusually formatted bit map.

There are 3 16-bit sections reserved, but only 4 bits are actually used
at this time. Those bits are:

3<<6	Enable use of the right mouse button as another CLEAR key
2<<6	Disable right mouse button as CLEAR
3<<4	Enable "key click" (a short audio tone) for the current window
2<<4	Disable key click on the current window

Since one of these options is global, it's mostly meant for config utils,
but since those config utils have to actually be written in something, we
gotta implement the call. :)