CC09: Disable some dead code by default
As it happens, I discovered that the compiler actually seems to still
contain all the code that the FLEX version used to allow using the Y
register as a variable, but which was disabled because OS-9 level 1 needs
a pointer to the program's data area. It appears to have been disabled by
setting the MAXREG symbol to 1 instead of 2, and leaving all the code that
dealt with Y as a variable dormant and unreachable.

In the long term, this means that some day a Level 2-only version of the
compiler might be built that is able to use two register vars, improving
speed and efficiency. It also means that it might be possible to recreate
the original FLEX version of the compiler. But in the short term, it means
that I can disable the code entirely and save some precious memory!