CC09: More "YREG" stuff
After some experiments with the code generator, I think we're getting
closer to the "original version" of the compiler when USE_YREG is enabled.

Two "register" variables are supported, and this variant of the compiler
seems to use absolute addressing for its variables like a FLEX system, or
like a machine language program in the BASIC environment.

Note that this is incompatible with the usual conventions we use for C
code on OS-9. No current compiled libraries can be used, and even 'cstart'
will need rewriting to make use of this variant on OS-9 level 2. It's also
not compatible AT ALL with level 1, so there's that.

But this might give some more flexibility for the future, so I think it's
worth looking into.