CPP: Save some RAM during runs
I noticed that quite a lot of preprocessor symbol replacement strings have
a significant number of trailing spaces after them, to an extent that in
aggregate, those spaces are wasting available memory in the symbol table.

So, let's do something about that, shall we?

The original 'savestr' function grabbed the rest of the current line using
strcpy (grab (strlen (string)), s); We could do that and change the length
of the resulting string, but because it's already allocated it's already
wasting memory we'd rather it didn't, so instead we want to figure out how
much to copy ahead of time.

So what we do is get the length of the string like usual, but after we do
that we check the last character of the length with isspace(), and if it
is, we reduce the length by 1 and try again. It's done with some...uhh,
interesting...pointer arithmetic, but it should be safe. and once we know
how long our final string is going to be, we allocate exactly as many
bytes as we need and terminate the string ourselves.

It's not going to affect most people, but people bumping up against the
limits should be happier.