CPP: some optimizations
Found some functions that weren't making appropriate use of 'register'
variables.

The best use of the 'register' storage class is usually for an often-used
variable, especially if that variable is used as an iterator in one or
more loops. A few functions didn't use any at all, and some just declared
that all of their variables should be held in registers -- which doesn't
help either, because we only have one register (sometimes two) available
to hold a variable, and the compiler picks the first one you tell it.

What you *want* to do is declare your iterators 'register' in order of
importance, such that you get the maximum benefit. For our compiler, you
can take advantage of the fact that some variables spend most of their
time in registers anyway (the result of a 'switch' expression gets held in
X for all the code inside the switch, so that can be a 'free' register
variable specific to switches).

Anyway, making the best use of register variables makes programs faster and
smaller! This commit saves several hundred bytes of code when compiled,
and should make the code run a bit faster as well.