KLibc: Add unsigned division, mod, and shr for the compiler to use
Unlike addition, subtraction and multiplication, unsigned division and
shifts to the right are different from the signed versions of the same.

For signed division/modulus, division has to do special processing to make
sure to do the right thing when one or both sides of the expression is
negative. With unsigned, because negative isn't a thing, we get to (and
in fact, must) avoid all that processing. The practical upshot is that
unsigned division of longs is a bit faster than signed.

For bit shifts, left is identical for both signed and unsigned, but in
signed right shifts the bit that gets shifted in will be a 1 if the
_previous_ leftmost bit was a 1. That is, shifting a negative number
to the right halves its magnitude but the number remains negative. That
is, 0b10000000 (0x80, -128) using "arithmetic shift right" becomes
0b11000000 (0xc0, -64), and repeating this just adds more 1 bits on the
left, pushing the existing bits rightward until you get 0xff, or -1.

For unsigned right shift, we need to use "logical shift right", where the
bits coming in from the left are always zeroes. 0b11111111 (0xff) is 255
in this form, and shifting rightward we want those numbers to be cut in
half too, so shifting rightward once we get 0b01111111 (0x7f, 127), and
so forth.

This adds _lumod, _ludiv, and _lushr special functions to the C library.