CC09: Support unsigned right shift and division in the compiler
Rather than add code to check for unsigned type to each operation, I
repeated the hack that McCosh used for the assignment operator tokens.
That is, I changed the order of the unsigned UDIV/UMOD/USHR tokens (which
are also used internally for operation tags in expression trees) to match
the order used for DIV/MOD/SHR, so that I can use a single set of checks
to turn a signed divide or shift into an unsigned one just by subtracting
the difference of two numbers.

Yes, it's a hack, but it's a clever one. And it's got comments, so readers
of the code should recognize it's something you want to be careful around.