CC09: Allow turning 'long' division into right shifts
I noticed something in the section of the 'long' expression translator,
in the section of a 'switch' for case TIMES (for multiplication):

 op = (op == TIMES) ? SHL : SHR;

This code section turns multiplications by a constant power of 2 into a
number of left shifts, for efficiency. But this (otherwise-useless) line
tells me it used to run for multiplication AND division, and at some point
somebody made it not do that any more.

Well, this re-enables the feature. Dividing a long (whether signed or not)
by a constant power of two will now use the appropriate shift operation
for speed.