CC09: change how single-character C variables are converted to asm symbols
All previous versions of the compiler would turn the single-character
variable name "a" into the assembly symbol "a_". I am pretty sure this was
done to prevent variables with the same names as 6809 registers, because
that causes real trouble.

The problem comes in because it means you can't have both a C variable
named "a" and one named "a_", they'll collide. This is bad, because both
of those names are individually legal.

This changes the code to instead prepend two underscores. The C standards
explicitly reserve all symbols with two leading underscores (or one, if
followed by an uppercase letter) for use by the implementation, so if you
create a symbol that conflicts, that's on you. :)