CPP: variadic macros from C99
This implements the '...' argument and __VA_ARGS__ symbol within macro
expansion.

A function-like macro may have a variable number of arguments if the last
named argument is '...'. Within the expansion, the final, unnamed, argument
is given the name __VA_ARGS__ and expands to the contents of any arguments
in excess of the number of named arguments.

That is, if you do:
#define eprintf(...) fprintf (stderr, __VA_ARGS__)

then the expansion of __VA_ARGS__ will contain a list of all the arguments
you supplied to the eprintf() macro.

Stringification should work on this as well, producing a string containing
the list of matching arguments.

This also introduces a workaround for a current limitation in cpp/dcpp:
Modern C differentiates between simple and function-like macros, in that
it can have two macros with the same name, differing in whether they take
arguments.

Well, we currently can't do that. When expanding a function-like macro that
doesn't supply an argument list, we now output a "no arguments" warning and
expand to the name of the macro. This should have the same effect as not
recognizing the macro at all, except of course for the warning. This used
to be an error.

Also increased the value of __DCC__ so users can detect support for
variadic macros.