CPP: simplify and regularize debug output
There were two copies of the "macro dump" code, and they were different
from one another (probably because there were two copies...).

This merges them into a new function printmac() in macros.c

The output looks like this:
[name] [argcount] ->text<-

In the output, the text of arguments is replaced by <#x>, where x is the
index of the argument.

Variadic function-like macros' argument count is followed by +, as in:

   eprintf  2+->fprintf(stderr, (<#1>),<#2)<-