CPP: findmac() doesn't need to be this way...
findmac() goes rather a long way to avoid strncmp(), saving and replacing
the last character of the string buffer with a nul byte to "make sure" it
can't overflow its buffer, and then the "delmac" (delete macro) function
actually works around this behavior by copying its argument into a
temporary buffer.

This is kinda dumb. What's worse is that it's got a bug where if it gets a
match, it fails to restore the modified byte in the passed-in buffer.

Rather than fix the bug in the hack, I choose to get rid of it by just
using strncmp like should have been done in the first place. :)

Also stripped out the workaround for the hack, both shrinking and speeding
up the binary.