CPP: rewrite function-like macro argument scanner
So I discovered that our argument scanner did not allow there to be an
empty last argument. To explain, for an (ANSI+) C preprocessor the macro
invocation "foo(a,)" is being supplied two arguments -- the first argument
is the string "a", and the second is the empty string "". Our argument
scanner was eating that last empty argument.

Along the way, I discovered that if we turned each argument string into a
valid null-terminated C string, we wouldn't need to go messing about with
counting characters and keeping both arguments and their lengths any more,
which means I can simplify copystr(), and so on.

The practical upshot is that dcpp is not just more standards-compliant, but
it's also faster and uses less memory too! :)