CPP: Another micro-optimization
Calling nxtch() will be (slightly) faster if its code is closer to that of
gch(), the function it calls. Why? Because calling nxtch() was always
going to be a long branch anyway, but if we move the code closer to gch(),
it will become a short branch instead. And it basically belongs in parse.c
anyway. :)