CPP: use the symbols for gch() calls
All the calls to gch() drive me a little nuts because I can never remember
which symbol has which value. For the record, the values are:
SKIPSP (0), gch() ignores comments and all non-newline whitespace.
KEEPSP (1), gch() ignores comments, but can result in cch being a space.

Anyway, let's not be passing bare numbers.