<iso10646.h>: New header
The ISO C standards add some "alternate spellings" for C operators that
aren't supported by the ISO 646 Basic Character Set, which is:

!"%&'()*+,-.0123456789:;<=>?
ABCDEFGHIJKLMNOPQRSTUVWXYZ
abcdefghijklmnopqrstuvwxyz

You may notice that some characters that are normally pretty critical in
the C language are missing from this list. You're right, and that can be
an issue for program portability. This header lets you replace the
logical OR operator, ||, by simply spelling it "or" instead.

These operators are required to be defined in the header <iso646.h>.