<sys/reg.h>: New header.
The header <sys/reg.h> is used to define the structure representing the
CPU's registers. It's primarily used for calls to the universal _os9()
system call, but it's also used to describe the format of the registers
stacked in an interrupt...and that format is actually different between
6809 and 6309 kernels because in native mode interrupts cause the E and
F registers to get stacked between the B and X registers, instead of
going B X with nothing between.

That's not very important for user code, because OS-9 system calls don't
use any of the 6309 registers, but if we want to be able to use C inside
system modules like drivers (and I for one definitely do), we need to be
able to understand the register stack and be able to find X, Y, etc.

Anyway, this creates a typedef REGISTERS that can be pointed to the
starting address of a "register stack", and accessed using normal struct
members. For convenience, there is a 16-bit member rg_d containing the
unsigned integer value of the combined D register.

To use REGISTERS in 6309 mode, instead of using the rg_* members, use
r3_*, as in addr->r3_y. Determining which mode to use is still left up to
you for now.