KLibc: time revamp
This splits up and updates the time functions, adds the canonical C time
conversion function mktime(), partially enables time zone functionality,
slices, dices, and makes julienne fries three different ways!

Previously, the 'daylight' and 'timezone' variables were forced to be
constant. Now they may be set to appropriate values, with timezone
representing an offset in seconds from GMT.

The system clock, the OS-9 setime/getime buffer 'struct sgtbuf', and C
broken-down time are all assumed to be set to local time, while time_t is
assumed to be in GMT. This means conversions between time_t and almost
anything else involve the timezone variable.

The mktime() function takes a C broken-down time structure 'struct tm'.
and returns the 'same' value as a time_t. If timezone is nonzero mktime()
will use it.

The cvtime() function does the same, except instead of 'struct tm', it
converts from an OS-9 'struct sgtbuf'. This function was previously named
o2utime(), for "OS-9 To Unix time", and may still be called by that name,
but since time_t has been defined by the C standards for 35 years now...

The localtime() function has been rewritten to use a more efficient
algorithm. It's smaller, and hopefully it's as fast as the assembly code
it replaces. If timezone is nonzero, localtime will use it.

Added a gmtime() function that functions like localtime() but does NOT
use the time zone. That is, if timezone is nonzero, localtime() and
gmtime() will give results differing by 'timezone' seconds. Care should
be taken not to use this to set the system clock for obvious reasons.

The function systime() converts from C broken-down time to an OS-9 time
buffer 'struct sgtbuf', with the only conversion being that the range of
the month field of 'sgtbuf' is 1-12, while that of 'tm' is 0-11. This
function was previously called u2otime(), and may still be called by that
name for backward compatibility, but it should be considered deprecated.

This commit is the first use of my new policy of trying to make the
assembly-optimized versions of functions optional. If the preprocessor
symbol ASMOPT is used (the default), the assembly versions of library
functions will be built, otherwise the C versions will be. My hope is that
this will make it easier to verify that the assembly is actually working
correctly.

Another new "policy" is that I'm trying to include test code into the
library source. For example, see localtime.c -- compiling this source code
with the TEST_MAIN symbol defined gives a program that tests the mktime,
localtime, gmtime, and asctime functions.