KLibc: Start working on CMOC+lwtools support
Pierre Sarrazin mentioned that the CMOC library is a bit lacking in normal
C functions, so I'm trying to make the source compile for both native OS-9
using DCC and the CMOC/lwtools cross-compilation environment.

To do that, I'm using a few tricks.

First, I wrote a couple of 'sed' scripts to massage the original source
code at a couple of stages to make it palatable to CMOC first, and lwasm
after it's gone through that.

dccmoc.sed converts McCosh-style #asm/#endasm preprocessor directives to
CMOC-style asm { } blocks. If there are more automatic changes I need to
make before preprocessing and compilation, they'll go in here.

rma2lw.sed does a couple of things -- it does a kinda reasonable job
translating RMA 'psect', 'csect', and 'vsect' directives into different
uses of lwasm's SECTION directive, it inserts EXPORT directives for all
global symbols (in RMA, these are the ones with a colon postfix), and it
adds duplicate exported symbols to hopefully be compatible with the way
CMOC generates labels.

That's a lot of confusing, but the practical upshot is that it turns
foo: pshs u

into
_foo EXPORT
foo EXPORT
_foo
foo pshs u

because CMOC needs an underscore prepended to all symbols meant to be
visible from C, and lwasm needs an EXPORT for each symbol to be visible
from other object files. We can't use the "export" pragma, because
then 'static' doesn't work right.

And for the files that we need to 'use' inside source files (such as
'os9defs.a' and 'stdio.a'), we have no choice but to modify those in place
so that they'll work right in either RMA or lwasm. To that end, I wrote a
macro to emulate 'csect' with 'SECTION _constants', which should handle
both ways csect is used.

I should note, this does not work yet. In collaboration with Pierre, I am
using a pre-release CMOC and made some changes myself to allow it to
compile C source files with suffixes other than .c, and even then CMOC
is not currently able to handle some of the functions in libc.