KLibc: Restore original getopt() source
The getopt() used in klibc is almost the same as the public domain getopt()
posted to the Usenet mod.sources newsgroup by Keith Bostic when he was at
Harvard, before he moved to Berkeley's Computer Systems Research Group to
build 4.4BSD.

The only difference is that someone replaced the 'tell' macro with a pair
of macros, tell and tell0, presumably due to a defect in c.prep. Well, we
don't use c.prep no mo. :)